elf: Add support to memory sealing
It uses the new Linux mseal syscall to seal memory mappings to avoid
further changes (such changing the memory protection or remap).  The
sealing is done in multiple places where the memory is supported to
be immuatable over program execution:

  * All DT_NEEDED dependencies from the binary, including the RELRO
    seguments after PT_GNU_RELRO setup.

  * The binary itself, including dynamic and static linked.  In both
    cases it is up either to binary or the loader to setup the
    sealing.

  * The ld.so.cache cache.

  * The vDSO segment (if existent).

  * Any preload libraries.

For binary dependencies, the RTLD_NODELETE is used to signal the
link_map should be sealed.  It also make dlopen objects with the
flag sealed as well.

The sealing is controlled by a new tunable, glibc.rtld.seal, with
three different states:

  0. Disabled  where no sealing is done.  This is the default.

  1. Enabled, where loader will issue the mseal syscall on the
     memory mappings but any failure will be ignored.

  2. Enforce, similar to Enabled but any failure from the mseal
     will terminate the process.

Checked on x86_64-linux-gnu.