elf: Add support to memory sealing for audit modules
It is done after the the loading process (instead at loading time
with RTLD_NODELETE), because the module might be unloaded due
nexistent or wrong la_version.  It also means that the la_version
and l_objopen/la_activety might be called without the object memory
sealed.

Checked on x86_64-linux-gnu.